Crate cycles_minting_canister

Structs

CyclesCanisterInitPayload

IcptsToCycles

SetAuthorizedSubnetworkListArgs
Argument taken by the set_authorized_subnetwork_list endpoint

Constants

CREATE_CANISTER_REFUND_FEE

DEFAULT_CYCLES_PER_XDR

MEMO_CREATE_CANISTER

MEMO_TOP_UP_CANISTER

TOP_UP_CANISTER_REFUND_FEE

Functions

create_canister_txn

top_up_canister_txn

Type Definitions

CreateCanisterResult
The result of create_canister transaction notification. In case of an error, contains the index of the refund block.

TopUpCanisterResult
The result of top_up_canister transaction notification. In case of an error, contains the index of the refund block.